backend/requirements.txt
fastapi==0.103.1
fastapi-mail==1.4.1
PyJWT==2.8.0
Requests==2.31.0
# SQLAlchemy==2.0.19
SQLAlchemy==1.4
Werkzeug==2.3.6
pydantic==2.3.0
pyyaml==6.0.1
uvicorn==0.23.2
psycopg2==2.9.9
logger
cachetools
python-jose

